+++ examples/threadpool_example.cpp
@@ -1,6 +1,6 @@
-﻿#include "core.h"
-#include "interface.h"
-#include "ithreadpool.h"
+#include "core/core.h"
+#include "core/interface.h"
+#include "threadpool/ithreadpool.h"
 #include "example_common.h"
 
 #include <atomic>