+++ plugins/logger/src/filelogger.h
@@ -3,7 +3,7 @@
  * @brief 基于 spdlog 的文件日志实现
  */
 #pragma once
-#include "ifilelogger.h"
+#include "logger/ifilelogger.h"
 #include <mutex>
 
 PCH_BEGIN_NAMESPACE